Add v1 of EthernetLock
Scans fingerprint. If enrolled print detected, makes a request to
server/validFingerprint.php?id={id}. At the moment, it just prints out
the HTTP response.

The PHP script contains an array of valid fingerprint IDs. If the ID
passed is in the array, a 200 OK response is issued. Otherwise a 403
Forbidden response is returned.